auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = ["fastapi (>=0.135.2,<0.136.0)", 
                "uvicorn (>=0.42.0,<0.43.0)", 
                "sqlalchemy (>=2.0.48,<3.0.0)", 
                "asyncpg (>=0.31.0,<0.32.0)", 
                "python-jose[cryptography] (>=3.5.0,<4.0.0)", 
                "passlib[bcrypt] (>=1.7.4,<2.0.0)", 
                "pydantic-settings (>=2.13.1,<3.0.0)", 
                "environs (>=14.6.0,<15.0.0)", 
                "pydantic (>=2.12.5,<3.0.0)", 
                "alembic (>=1.18.4,<2.0.0)", 
                "bcrypt (<4.1)",
                ]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.7,<0.16.0)"
]
